# docs/swagger.yaml
basePath: /
definitions:
  github_com_v-kuu_mini-marketplace_internal_model.Product:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  internal_http_api.CreateProductRequest:
    properties:
      name:
        type: string
      price:
        type: integer
    type: object
  internal_http_api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_http_api.PatchProductRequest:
    properties:
      name:
        type: string
      price:
        type: integer
    type: object
  internal_http_api.UpdateProductRequest:
    properties:
      name:
        type: string
      price:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: A small CRUD api serving a simple Product model
  title: mini-marketplace
  version: "1.0"
paths:
  /products:
    get:
      description: Returns all products in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_v-kuu_mini-marketplace_internal_model.Product'
            type: array
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
      summary: Get products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product with idempotency
      parameters:
      - description: Product to create
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_api.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_v-kuu_mini-marketplace_internal_model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    get:
      description: Returns a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_v-kuu_mini-marketplace_internal_model.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
      summary: Get a product by ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Partially updates a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_api.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_v-kuu_mini-marketplace_internal_model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
      summary: Patch a product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates an existing product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_api.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_v-kuu_mini-marketplace_internal_model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http_api.ErrorResponse'
      summary: Update a product
      tags:
      - products
swagger: "2.0"